Updated FSM test. Started to impl single entry tests
